+++ src/dotnet/MLSearch.Service/Engine/OpenSearch/Configuration/PlainIndexSettingsFactory.cs
@@ -1,0 +1,18 @@
+
+using ActualChat.MLSearch.Engine.OpenSearch.Setup;
+using Microsoft.Extensions.Options;
+
+namespace ActualChat.MLSearch.Engine.OpenSearch.Configuration;
+
+internal sealed class PlainIndexSettingsFactory(IndexNames indexNames, IClusterSetup clusterSetup)
+    : IOptionsFactory<PlainIndexSettings>
+{
+    public PlainIndexSettings Create(string name)
+    {
+        ArgumentException.ThrowIfNullOrEmpty(name);
+
+        var clusterSettings = clusterSetup.Result;
+        var indexName = indexNames.GetFullName(name, clusterSettings);
+        return new PlainIndexSettings(indexName);
+    }
+}
