+++ src/dotnet/MLSearch.Service/Engine/OpenSearch/Configuration/SemanticIndexSettingsFactory.cs
@@ -2,20 +2,7 @@
 using ActualChat.MLSearch.Engine.OpenSearch.Setup;
 using Microsoft.Extensions.Options;
 
-namespace ActualChat.MLSearch.Engine.OpenSearch;
-
-internal sealed class PlainIndexSettingsFactory(IndexNames indexNames, IClusterSetup clusterSetup)
-    : IOptionsFactory<PlainIndexSettings>
-{
-    public PlainIndexSettings Create(string name)
-    {
-        ArgumentException.ThrowIfNullOrEmpty(name);
-
-        var clusterSettings = clusterSetup.Result;
-        var indexName = indexNames.GetFullName(name, clusterSettings);
-        return new PlainIndexSettings(indexName);
-    }
-}
+namespace ActualChat.MLSearch.Engine.OpenSearch.Configuration;
 
 internal sealed class SemanticIndexSettingsFactory(IndexNames indexNames, IClusterSetup clusterSetup)
     : IOptionsFactory<SemanticIndexSettings>
